# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TP2 API de films",
        "description": "API REST pour l’authentification TP2.1",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/films": {
            "post": {
                "tags": [
                    "Films"
                ],
                "summary": "Créer un film (admin seulement)",
                "operationId": "ce1501bac824524d1e4706075650bf11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "release_year",
                                    "length",
                                    "description",
                                    "rating",
                                    "special_features",
                                    "image",
                                    "language_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 50
                                    },
                                    "release_year": {
                                        "type": "integer",
                                        "example": 2024
                                    },
                                    "length": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "rating": {
                                        "type": "string",
                                        "maxLength": 5
                                    },
                                    "special_features": {
                                        "type": "string",
                                        "maxLength": 200
                                    },
                                    "image": {
                                        "type": "string",
                                        "maxLength": 40
                                    },
                                    "language_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Film créé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/films/{id}": {
            "put": {
                "tags": [
                    "Films"
                ],
                "summary": "Modifier un film (admin seulement)",
                "operationId": "d2fbe78b6e79c3f1cfb5067d25f9a039",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Film"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Film modifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Films"
                ],
                "summary": "Supprimer un film (admin seulement)",
                "operationId": "decacbac55d9f088563f35def674b92a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Film supprimé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Voir ses propres informations",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Utilisateur trouvé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Utilisateur introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/{id}/password": {
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Mettre à jour son mot de passe",
                "operationId": "4f86d763349c89bca0a239bd4334c7b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "example": "newPassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newPassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mot de passe mis à jour"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Film": {
                "title": "Film",
                "description": "Film model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "ACADEMY DINOSAUR"
                    },
                    "description": {
                        "type": "string",
                        "example": "A Epic Drama of a Feminist And a Mad Scientist who must Battle a Teacher in The Canadian Rockies"
                    },
                    "release_year": {
                        "type": "integer",
                        "example": 2006
                    },
                    "language_id": {
                        "description": "Foreign key referring to Language",
                        "type": "integer",
                        "example": 1
                    },
                    "length": {
                        "description": "Film length in minutes",
                        "type": "integer",
                        "example": 86
                    },
                    "rating": {
                        "type": "string",
                        "example": "PG"
                    },
                    "special_features": {
                        "type": "string",
                        "example": "Deleted Scenes,Behind the Scenes"
                    },
                    "image": {
                        "type": "string",
                        "example": ""
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2006-02-15 15:03:42"
                    }
                },
                "type": "object"
            }
        }
    },
...
}